fix: moving package versions to Directory.Packages.props from csproj in init command
When nunit tests project is created with dotnet new then nugets with versions are added to csproj (and restored) instead of Directory.Packages.props (this file is created later in the process, but even if it's present, nugets are put fully in csproj - it's probably dotnet cli bug). So I need to move versions to Directory.Packages.props.

+++ tests/xpd.tests/IntegrationTests/InitIntegrationTests.cs
@@ -87,11 +87,48 @@ public class InitIntegrationTests : InitTestsBase
 
     private static string PrepareOutputDir()
     {
-        const string outputDir = "TestOutputDir";
-        var assemblyLocation = typeof(InitIntegrationTests).Assembly.Location;
-        var outputPath = Path.Combine(new FileInfo(assemblyLocation).DirectoryName!, outputDir);
-        Directory.Delete(outputPath, true);
+        const string outputDir = "XpdIntegrationTestsOutputDir";
+        var rootRepoFolder = GetRootRepoFolder();
+        var outputPath = Path.Combine(rootRepoFolder, "..", outputDir);
+        if (Directory.Exists(outputPath))
+        {
+            Directory.Delete(outputPath, true);
+        }
+
         return outputPath;
+    }
+
+    private static string GetRootRepoFolder()
+    {
+        var currentDir = Directory.GetCurrentDirectory();
+        while (currentDir is not null && !HasGitFolder(currentDir))
+        {
+            currentDir = Directory.GetParent(currentDir)?.FullName;
+        }
+
+        if (currentDir is null)
+        {
+            throw new Exception("Could not find root repo folder.");
+        }
+
+        // Make additional checks for common files to make sure it is root repo folder
+        var packagesPropsExists = new FileInfo(
+            Path.Combine(currentDir, "Directory.Packages.props")
+        ).Exists;
+        var solutionExists = new DirectoryInfo(currentDir).GetFiles("*.sln").Length == 1;
+        var solutionExistsInSrc =
+            new DirectoryInfo(Path.Combine(currentDir, "src")).GetFiles("*.sln").Length == 1;
+
+        if (!packagesPropsExists && !(solutionExists || solutionExistsInSrc))
+        {
+            throw new Exception("Root repo folder doesn't have required files.");
+        }
+
+        Console.WriteLine($"Root repo folder: {currentDir}");
+        return currentDir;
+
+        static bool HasGitFolder(string folder) =>
+            Directory.EnumerateFileSystemEntries(folder).Any(f => f.EndsWith(".git"));
     }
 
     public class DotnetToolsManifest
